requirements.txt
flet==0.82.2
Pillow>=10.0.0
pyinstaller>=6.0.0